realtime subpackage refactor
