+++ pycc/rt/lasers.py
